+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# 1.0.0-x
+# 1.1.0-x
+
 ## Improvements
 
 - remove the fixed mysql socket location
@@ -6,6 +7,17 @@
 - base template has been moved from `app/Resources/views` to `PimEnrichBundle/Resources/views`
 
 ## Bug fixes
+
+## BC breaks
+- `PimEnterprise\Bundle\WorkflowBundle\DependencyInjection\Compiler\ResolveDoctrineOrmTargetEntitiesPass` has been renamed to `ResolveDoctrineTargetModelsPass`
+
+# 1.0.0-x
+
+## Improvements
+- Setup relationships of published products with interfaces in order to easy overriding
+
+## Bug fixes
+
 ## BC breaks
 
 # 1.0.0